ufff

+++ src/main/java/ru/kata/spring/boot_security/demo/model/Role.java
@@ -13,7 +13,7 @@ import java.util.Set;
 public class Role implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String name;
     @ManyToMany(mappedBy = "roles")
     private Set<User> users = new HashSet<>();
@@ -21,20 +21,20 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public Role(Long id) {
+    public Role(int id) {
         this.id = id;
     }
 
-    public Role(Long id, String name) {
+    public Role(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
